Move trigger refresh logic to token handler
Before, a userattribute handler would check if the ExtraByProvider
field was null and trigger a refresh. ExtraByProvider contains
data that is available only on tokens. When there was no token
associated with the user corresponding to the userattribute, the
handler would fire endlessly. This happened because the field
continued to be null but the refresh process would update the
LastRefreshed field which would cause the userattribute handler
to execute again. Now, the ExtraByProvider check is done on a
token handler. The token handler is aware that the tokens
existence may make populating the field possible and the refresh
process, makes no changes to the token object. This avoids nonstop
firing of handlers.