Fix service account token secret backoff timing
Without this change, if a user is created, is added to a project or
cluster, and then makes a request in rapid succession (as in a CI case),
it is highly likely that the PRTB or CRTB controller will not have
finished adding the user's rolebindings to the project or cluster in
between steps 2 and 3. The cause is the refactoring of the service
account token generator in 37c7ee19 which added a wait to the service
account creator to ensure the token was populated before returning. The
wait loop would check the current secret object, then refresh the
secret, then wait 2 seconds before checking that secret again. The
original secret object was certainly not populated, and the first
refresh was only a few nanoseconds afterward and so the first-time
refreshed secret was also almost certainly not populated, so it was not
until the second refresh that the secret was populated and ready. With
the loop timing set to 2 seconds, this meant the wait took a full 4
seconds.

This change reorders the wait loop to refresh the secret first thing, to
avoid an extra loop, and reduces the step period to 2 milliseconds. This
is enough time for the token to populate on the 2nd or 3rd retry and
makes it much more likely the controller can finish setting up
rolebindings before the user needs them.