Assign special error type to errors related to incompatible template versions
This special error type (`IncompatibleTemplateVersionErr`) can be used by other packages to identify whether the error returned by a call to one of the Catalog Manager's methods is tied to an real error with trying to execute a function call or the result of a successful call that returns that nothing is available to install.

In most cases, the consumer of this error will just emit the error and move on, but in certain cases (such as in `pkg/controllers/managementuserlegacy/systemimage/systemimage.go`), the fact that there is no compatible version results in a different action that needs to be taken (i.e. silently ignore and move on)