Add integration tests for the Steve API
The Steve API is essentially a mirror of the Kubernetes API with some
subtle differences. This commit adds integration tests to validate the
steve API, especially the resource listing API.

To accomplish this, some framework changes were needed:

- Since Steve is completely separate from Norman, and Steve's API is
  diverging from Norman's, it's not appropriate to use Norman types as
  parameters in Steve clients. The norman types are swapped for net/url
  query values, which is more similar to how a generic client (like a
  Javascript dashboard) will be using the Steve API.
- A NamespacedSteveClient object is added to access endpoints under
  /v1/{resource}/{namespace} rather than filtering with the
  fieldSelector parameter.
- Helpers are added to manage Roles and RoleBindings through kubeapi
  rather than Norman.
- A default cluster name is set during setup so that integration tests
  can use it. Previously, only the validation tests could use the config
  cluster name which was set up in the Jenkinsfile.

The new tests use secrets as the sample resources since they are quick
to create.

Basic CRUD operations using global and namespaced resource endpoints are
tested.

The new list tests use three users to test three main scenarios: a
project owner, a user with access for one namespace, and a user with
access for a few resources in a few namespaces. This tests Steve's
ability to collect and return results across multiple access segments.

The new tests check the responses for basic Kubernetes parameters,
fieldSelector, labelSelector, limit, and continue, for all three user
scenarios, with or without a namespace selected.